Reorganize code to remove jest dependency in test_avm.ts
