+++ yarn-project/bb-prover/src/test/test_avm.ts
@@ -1,10 +1,7 @@
 import {
-  AvmCircuitInputs,
-  AvmVerificationKeyData,
   AztecAddress,
   ContractStorageRead,
   ContractStorageUpdateRequest,
-  FunctionSelector,
   Gas,
   GlobalVariables,
   Header,
@@ -31,36 +28,10 @@ import {
 } from '@aztec/circuits.js';
 import { computeVarArgsHash } from '@aztec/circuits.js/hash';
 import { padArrayEnd } from '@aztec/foundation/collection';
-import { Fr } from '@aztec/foundation/fields';
-import { createDebugLogger } from '@aztec/foundation/log';
-import { AvmSimulator, type PublicContractsDB, type PublicExecutionResult, type PublicStateDB } from '@aztec/simulator';
-import { PublicSideEffectTrace } from '@aztec/simulator';
-import {
-  getAvmTestContractBytecode,
-  initContext,
-  initExecutionEnvironment,
-  initHostStorage,
-  initPersistableStateManager,
-  resolveAvmTestContractAssertionMessage,
-} from '@aztec/simulator/avm/fixtures';
-import { SerializableContractInstance } from '@aztec/types/contracts';
-
-import fs from 'node:fs/promises';
-import { tmpdir } from 'node:os';
-import path from 'path';
-
-import { type BBSuccess, BB_RESULT, generateAvmProof, verifyAvmProof } from '../bb/execute.js';
-import { extractAvmVkData } from '../verification_key/verification_key_data.js';
-
-// The paths for the barretenberg binary and the write path are hardcoded for now.
-const bbPath = path.resolve('../../barretenberg/cpp/build/bin/bb');
-const bbWorkingDirectory = await fs.mkdtemp(path.join(tmpdir(), 'bb-'));
-
-const internalLogger = createDebugLogger('aztec:avm-proving-test');
-const logger = (msg: string, _data?: any) => internalLogger.verbose(msg);
+import { type PublicExecutionResult } from '@aztec/simulator';
 
 // TODO: pub somewhere more usable - copied from abstract phase manager
-const getPublicInputs = (result: PublicExecutionResult): PublicCircuitPublicInputs => {
+export function getPublicInputs(result: PublicExecutionResult): PublicCircuitPublicInputs {
   return PublicCircuitPublicInputs.from({
     callContext: result.executionRequest.callContext,
     proverAddress: AztecAddress.ZERO,
@@ -111,119 +82,4 @@ const getPublicInputs = (result: PublicExecutionResult): PublicCircuitPublicInpu
     // TODO(@just-mitch): need better mapping from simulator to revert code.
     revertCode: result.reverted ? RevertCode.APP_LOGIC_REVERTED : RevertCode.OK,
   });
-};
-
-export async function proveAvmTestContract(
-  // NOTE(AD): these are hackishly passed instead of
-  // importing them as this module gets exported in our index.ts and jest is only a dev dependency/
-  // In the future, a testing helper package that depends on jest might be a better solution
-  jest: any,
-  jestMock: any,
-  functionName: string,
-  calldata: Fr[] = [],
-  assertionErrString?: string,
-  globalVariables?: GlobalVariables,
-): Promise<BBSuccess> {
-  const startSideEffectCounter = 0;
-  const functionSelector = FunctionSelector.random();
-  const globals = globalVariables == undefined ? GlobalVariables.empty() : globalVariables;
-  const environment = initExecutionEnvironment({ functionSelector, calldata, globals });
-
-  const contractsDb = jestMock() as PublicContractsDB;
-  const contractInstance = new SerializableContractInstance({
-    version: 1,
-    salt: new Fr(0x123),
-    deployer: new Fr(0x456),
-    contractClassId: new Fr(0x789),
-    initializationHash: new Fr(0x101112),
-    publicKeysHash: new Fr(0x161718),
-  }).withAddress(environment.address);
-  (contractsDb.getContractInstance as any).mockResolvedValue(Promise.resolve(contractInstance));
-
-  const storageDb = jestMock() as PublicStateDB;
-  const storageValue = new Fr(5);
-  (storageDb.storageRead as any).mockResolvedValue(Promise.resolve(storageValue));
-
-  const hostStorage = initHostStorage({ contractsDb });
-  const trace = new PublicSideEffectTrace(startSideEffectCounter);
-  const persistableState = initPersistableStateManager({ hostStorage, trace });
-  const context = initContext({ env: environment, persistableState });
-  const nestedCallBytecode = getAvmTestContractBytecode('add_args_return');
-  jest.spyOn(hostStorage.contractsDb, 'getBytecode').mockResolvedValue(nestedCallBytecode);
-
-  const startGas = new Gas(context.machineState.gasLeft.daGas, context.machineState.gasLeft.l2Gas);
-
-  // Use a simple contract that emits a side effect
-  const bytecode = getAvmTestContractBytecode(functionName);
-
-  // First we simulate (though it's not needed in this simple case).
-  const simulator = new AvmSimulator(context);
-  const avmResult = await simulator.executeBytecode(bytecode);
-
-  if (assertionErrString == undefined) {
-    expect(avmResult.reverted).toBe(false);
-  } else {
-    // Explicit revert when an assertion failed.
-    expect(avmResult.reverted).toBe(true);
-    expect(avmResult.revertReason).toBeDefined();
-    expect(resolveAvmTestContractAssertionMessage(functionName, avmResult.revertReason!)).toContain(assertionErrString);
-  }
-
-  const pxResult = trace.toPublicExecutionResult(
-    environment,
-    startGas,
-    /*endGasLeft=*/ Gas.from(context.machineState.gasLeft),
-    /*bytecode=*/ simulator.getBytecode()!,
-    avmResult,
-    functionName,
-  );
-
-  const avmCircuitInputs = new AvmCircuitInputs(
-    functionName,
-    /*bytecode=*/ simulator.getBytecode()!, // uncompressed bytecode
-    /*calldata=*/ context.environment.calldata,
-    /*publicInputs=*/ getPublicInputs(pxResult),
-    /*avmHints=*/ pxResult.avmCircuitHints,
-  );
-
-  // Then we prove.
-  const proofRes = await generateAvmProof(bbPath, bbWorkingDirectory, avmCircuitInputs, logger);
-  expect(proofRes.status).toEqual(BB_RESULT.SUCCESS);
-
-  return proofRes as BBSuccess;
-}
-
-/**
- * If assertionErrString is set, we expect a (non exceptional halting) revert due to a failing assertion and
- * we check that the revert reason error contains this string. However, the circuit must correctly prove the
- * execution.
- */
-export async function proveAndVerifyAvmTestContract(
-  // NOTE(AD): these are hackishly passed instead of
-  // importing them as this module gets exported in our index.ts and jest is only a dev dependency/
-  // In the future, a testing helper package that depends on jest might be a better solution
-  jest: any,
-  jestMock: any,
-  functionName: string,
-  calldata: Fr[] = [],
-  assertionErrString?: string,
-  globalVariables?: GlobalVariables,
-) {
-  const succeededRes = await proveAvmTestContract(
-    jest,
-    jestMock,
-    functionName,
-    calldata,
-    assertionErrString,
-    globalVariables,
-  );
-
-  // Then we test VK extraction and serialization.
-  const vkData = await extractAvmVkData(succeededRes.vkPath!);
-  AvmVerificationKeyData.fromBuffer(vkData.toBuffer());
-
-  // Then we verify.
-  const rawVkPath = path.join(succeededRes.vkPath!, 'vk');
-  const verificationRes = await verifyAvmProof(bbPath, succeededRes.proofPath!, rawVkPath, logger);
-  expect(verificationRes.status).toBe(BB_RESULT.SUCCESS);
 }
