+++ yarn-project/ivc-integration/package.json
@@ -66,6 +66,7 @@
   },
   "devDependencies": {
     "@aztec/bb-prover": "workspace:^",
+    "@aztec/simulator": "workspace:^",
     "@jest/globals": "^29.5.0",
     "@msgpack/msgpack": "^3.0.0-beta2",
     "@types/jest": "^29.5.0",
