+++ yarn-project/ivc-integration/src/avm_integration.test.ts
@@ -1,5 +1,12 @@
-import { type BBSuccess, BB_RESULT, generateProof, verifyProof } from '@aztec/bb-prover';
-import { proveAvmTestContract } from '@aztec/bb-prover';
+import {
+  type BBSuccess,
+  BB_RESULT,
+  generateAvmProof,
+  generateProof,
+  getPublicInputs,
+  verifyProof,
+} from '@aztec/bb-prover';
+import { AvmCircuitInputs, FunctionSelector, Gas, GlobalVariables } from '@aztec/circuits.js';
 import {
   AVM_PROOF_LENGTH_IN_FIELDS,
   AVM_VERIFICATION_KEY_LENGTH_IN_FIELDS,
@@ -9,10 +16,21 @@ import { Fr } from '@aztec/foundation/fields';
 import { createDebugLogger } from '@aztec/foundation/log';
 import { BufferReader } from '@aztec/foundation/serialize';
 import { type FixedLengthArray } from '@aztec/noir-protocol-circuits-types/types';
+import { AvmSimulator, type PublicContractsDB, PublicSideEffectTrace, type PublicStateDB } from '@aztec/simulator';
+import {
+  getAvmTestContractBytecode,
+  initContext,
+  initExecutionEnvironment,
+  initHostStorage,
+  initPersistableStateManager,
+  resolveAvmTestContractAssertionMessage,
+} from '@aztec/simulator/avm/fixtures';
+import { SerializableContractInstance } from '@aztec/types/contracts';
 
 import { jest } from '@jest/globals';
 import fs from 'fs/promises';
 import { mock } from 'jest-mock-extended';
+import { tmpdir } from 'node:os';
 import os from 'os';
 import path from 'path';
 import { fileURLToPath } from 'url';
@@ -24,9 +42,9 @@ import { MockPublicKernelCircuit, witnessGenMockPublicKernelCircuit } from './in
 
 jest.setTimeout(120_000);
 
-describe('AVM Integration', () => {
-  const logger = createDebugLogger('aztec:avm-integration');
+const logger = createDebugLogger('aztec:avm-integration');
 
+describe('AVM Integration', () => {
   let bbWorkingDirectory: string;
   let bbBinaryPath: string;
 
@@ -55,7 +73,7 @@ describe('AVM Integration', () => {
   }
 
   it('Should generate and verify an ultra honk proof from an AVM verification', async () => {
-    const bbSuccess = await proveAvmTestContract(jest, mock, 'new_note_hash', [new Fr(1)]);
+    const bbSuccess = await proveAvmTestContract('new_note_hash', [new Fr(1)]);
 
     const avmProofPath = bbSuccess.proofPath;
     const avmVkPath = bbSuccess.vkPath;
@@ -104,3 +122,81 @@ describe('AVM Integration', () => {
     expect(verifyResult.status).toBe(BB_RESULT.SUCCESS);
   });
 });
+
+// Helper
+
+const proveAvmTestContract = async (
+  functionName: string,
+  calldata: Fr[] = [],
+  assertionErrString?: string,
+): Promise<BBSuccess> => {
+  const startSideEffectCounter = 0;
+  const functionSelector = FunctionSelector.random();
+  const globals = GlobalVariables.empty();
+  const environment = initExecutionEnvironment({ functionSelector, calldata, globals });
+
+  const contractsDb = mock<PublicContractsDB>();
+  const contractInstance = new SerializableContractInstance({
+    version: 1,
+    salt: new Fr(0x123),
+    deployer: new Fr(0x456),
+    contractClassId: new Fr(0x789),
+    initializationHash: new Fr(0x101112),
+    publicKeysHash: new Fr(0x161718),
+  }).withAddress(environment.address);
+  contractsDb.getContractInstance.mockResolvedValue(await Promise.resolve(contractInstance));
+
+  const storageDb = mock<PublicStateDB>();
+  const storageValue = new Fr(5);
+  storageDb.storageRead.mockResolvedValue(await Promise.resolve(storageValue));
+
+  const hostStorage = initHostStorage({ contractsDb });
+  const trace = new PublicSideEffectTrace(startSideEffectCounter);
+  const persistableState = initPersistableStateManager({ hostStorage, trace });
+  const context = initContext({ env: environment, persistableState });
+  const nestedCallBytecode = getAvmTestContractBytecode('add_args_return');
+  jest.spyOn(hostStorage.contractsDb, 'getBytecode').mockResolvedValue(nestedCallBytecode);
+
+  const startGas = new Gas(context.machineState.gasLeft.daGas, context.machineState.gasLeft.l2Gas);
+
+  // Use a simple contract that emits a side effect
+  const bytecode = getAvmTestContractBytecode(functionName);
+  // The paths for the barretenberg binary and the write path are hardcoded for now.
+  const bbPath = path.resolve('../../barretenberg/cpp/build/bin/bb');
+  const bbWorkingDirectory = await fs.mkdtemp(path.join(tmpdir(), 'bb-'));
+  // First we simulate (though it's not needed in this simple case).
+  const simulator = new AvmSimulator(context);
+  const avmResult = await simulator.executeBytecode(bytecode);
+
+  if (assertionErrString == undefined) {
+    expect(avmResult.reverted).toBe(false);
+  } else {
+    // Explicit revert when an assertion failed.
+    expect(avmResult.reverted).toBe(true);
+    expect(avmResult.revertReason).toBeDefined();
+    expect(resolveAvmTestContractAssertionMessage(functionName, avmResult.revertReason!)).toContain(assertionErrString);
+  }
+
+  const pxResult = trace.toPublicExecutionResult(
+    environment,
+    startGas,
+    /*endGasLeft=*/ Gas.from(context.machineState.gasLeft),
+    /*bytecode=*/ simulator.getBytecode()!,
+    avmResult,
+    functionName,
+  );
+
+  const avmCircuitInputs = new AvmCircuitInputs(
+    functionName,
+    /*bytecode=*/ simulator.getBytecode()!, // uncompressed bytecode
+    /*calldata=*/ context.environment.calldata,
+    /*publicInputs=*/ getPublicInputs(pxResult),
+    /*avmHints=*/ pxResult.avmCircuitHints,
+  );
+
+  // Then we prove.
+  const proofRes = await generateAvmProof(bbPath, bbWorkingDirectory, avmCircuitInputs, logger.info);
+  expect(proofRes.status).toEqual(BB_RESULT.SUCCESS);
+
+  return proofRes as BBSuccess;
+};
